process empty samples

+++ wrappers/fastq_mv/script.py
@@ -5,7 +5,10 @@ import os
 from snakemake.shell import shell
 
 for in_file, out_file in zip(snakemake.params.fastqs_in, snakemake.output.fastqs_out):
-    shell("mv " + in_file + " " + out_file)
+    if os.path.isfile(in_file):
+        shell("mv " + in_file + " " + out_file)
+    else:
+        shell("touch " + out_file)
     if os.path.isfile(in_file.replace("_R1_001.fastq.gz","_R2_001.fastq.gz")):
         shell("mv " + in_file.replace("_R1_001.fastq.gz","_R2_001.fastq.gz") + " " + out_file.replace("_R1_001.fastq.gz","_R2_001.fastq.gz"))
     if os.path.isfile(in_file.replace("_R1_001.fastq.gz","_R3_001.fastq.gz")):
